🚚 [move] moclue 삭제 common으로 이동
